requirements.txt
google-generativeai
streamlit
PyPDF2
pandas
python-dotenv